[Driver][FPGA] Improve output project folder behaviors for multi-file
When performing multi-file compilations with AOT for FPGA, there is an
additional output report folder that is created with the separate aoc
call.  This folder was continually over-written due to each subsequent
aoc call creating the same output folder name.

Add an additional capability to llvm-foreach:
let the user provide an --out-increment='file' argument. This will append
the number representing the current call made by llvm-foreach to the
given 'file' name.