[SYCL] Ensure the deferred diagnostics correctly handles discarded stmts
The deferred diagnostic 'checkFunc' handling uses the StmtVisitor to
check through all the calls. This inadvertently ends up diagnosing
discarded statements (that is, the inactive branch of the constexpr-if).
This patch adds a handler to only visit the active side.

We don't have to worry about visiting the conditional(that is, the
'rest' of the children) because we don't care about actions that happen
in the constant-expression, things that happen in a constexpr context
are typically allowed.