[SYCL] Temporary use host allocations for multi-device contexts
Currently Level Zero plugin allocates all buffers on a device with
index 0. If a context has multiple devices this allocation used on all
devices which results in an error. Temporary make Level Zero plugin to
use host allocations for multi-device contexts because such allocations
are accessible by all devices. In case of a single discrete device or
if context consist of sub-devices of a same device we still can use
device allocation.