[SYCL] Add is_device_copyable type trait
Currently the default behavior of is_device_copyable<T> is equal to:

(is_trivially_copyable_v<T> ||
 (is_trivially_copy_constructible_v<T> && is_trivially_destructible_v<T>))

That is done for compatibility with the current SYCL behavior that allowed
passing trivially copy-constructible/destructible types between devices.
The warning for non-trivially copyable but constructible/destructible types
is implemented but is not printed, which will be fixed separately.

Specialization of is_device_copyable type trait is added only for
std::tuple - which is one of so called 'implicitly device copyble' types.
That is done as example of how it is going to be done for the rest of
'implicitly device copyable' types mentioned in SYCL-2020 SPEC.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>