[Driver][SYCL] Integration footer support (#3568)
Similar to the integration header, the device compilation can produce
an integration footer.  Have the driver take this generated footer and
add it to the host compilation.

The host source file is first preprocessed, then the integration footer is
appended to the host source file then compiled.

Contains the modifications to incorporate the integration footer under
and interim switch -fsycl-use-footer.  The plan is to enable this functionality
by default.

New tool named 'append-file' created which appends 2 files together.  This
is integrated into the toolchain to take the preprocessed host compilation
and appends the generated integration footer before doing another full
host compilation against that appended file.