[SYCL] Improve DPC++ runtime performance running on L0 (#3797)
Fixes a performance issue due to https://github.com/intel/llvm/pull/3612

This change does two things to fix the performance issue.
The first is in the level zero plugin.  Here the change is to only
close and submit the batch if the event being queried is one that
will be signalled one of the commands in the batch.

The second change is in the sycl run-time itself.  This change
prevents the event cleanup code from querying every single event
that is outstanding in the system.  This is necessary to prevent
the most recent events (which are likely to be in the plug-ins batch)
from being queried.

Both changes are required to fix the performance issue.