Mark compiler ITT wrappers with always_inline.
Compiler generated ITT annotations may interfere with debugging
the instrumented user code even in non-ITT mode (when ITT is disabled
via the specialization constant). This change marks the wrappers
used by the compiler generated code with always_inline so that
device compiler are able to get rid of the extra code in non-ITT mode.

Signed-off-by: Vyacheslav Zakharin <vyacheslav.p.zakharin@intel.com>