Implement SPV_INTEL_fpga_invocation_pipelining_attributes extension
Some FPGA devices and toolchains can support customizable levels or
implementation of pipeline parallelism when mapping a SPIR-V module
to hardware. Through pipeline parallelism, multiple invocations of
a kernel or function can execute concurrently.

This extension adds decorations to request that a kernel or function
support invocations at a specified initiation interval, that multiple
invocations are forbidden from executing concurrently, or that the
kernel or function is limited to a maximum number of concurrent
invocations.

Spec:
https://github.com/KhronosGroup/SPIRV-Registry/blob/38a7e56b9b7feab7fa8b3bea08770f7232b76ed2/extensions/INTEL/SPV_INTEL_fpga_invocation_pipelining_attributes.asciidoc

Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/e303a5f