[Driver][SYCL] Do not imply defaultlib msvcrt for non-MSVC Windows
When compiling for Windows target, we should only be adding the
defaultlib msvcrt directive when performing MSVC compatible compilations.
If adding msvcrt for Linux style compilations, we do not know if
the library should be the regular library or the debug library, as that
is determined by the use of /MD or /MDd.