[SYCL] Refactor stream class handing implementation (#3646)
Currently the stream class when passed from host to device requires additional handling.
It is handled as a wrapper struct for several accessors. In order to create the AST, the stream class is processed as a whole and each accessor is processed individually.
This patch is to simplify the process. An __init and __finalize methods are introduced that eliminate the need to process the field individually. The steps are as follows:

1.	Look up for the member function sycl::stream::__init
2.	Copy the sycl::stream::__init arguments into the kernel function argument list
3.	Allocate a sycl::stream object and call sycl::stream::__init with the kernel argument
4.	Call sycl::stream::__finalize with the kernel argument at the end of the kernel

Signed-off-by: Zahira Ammarguellat <zahira.ammarguellat@intel.com>