Use SPIRVBinaryInst for OpFMod and OpSMod
OpFMod and OpSMod were implemented using a separate class, but
seemingly without an underlying motivation.  Instead of a separate
class, reuse the SPIRVBinaryInst class to unify OpFMod and OpSMod
handling with other binary instructions.

This change enables supporting FMod and SMod operators with
OpSpecConstantOp, which will be done in a future commit.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/b3b6819