Handle OpSpecConstantOp with VectorShuffle
Modify the `SPIRVShuffleVector` constructor to allow passing a nullptr
basic block (as is the case for variable initializers).

Modify the `SPIRVShuffleVector` constructor to take `SPIRVId`s instead
of `SPIRVValue`s, which better reflects what is stored in the class,
and saves us an unnecessary ID-to-Value-to-ID round trip in
`createInstFromSpecConstantOp`.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/72f99e3