Handle OpSpecConstantOp with CompositeExtract and CompositeInsert
Modify the constructors to allow passing a nullptr basic block (as is
the case for variable initializers).

Modify the constructors to take `SPIRVId`s instead of `SPIRVValue`s,
which better reflects what is stored in the class, and saves us an
unnecessary ID-to-Value-to-ID round trip in
`createInstFromSpecConstantOp`.

Modify test/opundef.spt to return the constructed value, so that it
does not get optimized out.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/9bc1d5a