[SYCL] Force instantiation of parallel_for before use of KernelInfo
With the new __builtin_sycl_unique_stable_name implementation, we need
to make sure we instantiate the kernel call BEFORE we use the builtin
via KernelInfo. We can do this by naming the function (but only if it is
a static/global, so this also changes those functions to static).

Additionally, we need to make sure we do this in the same way that
parallel_for_wrapper does, otherwise we'll end up with 2 kernels
instantiated with the same name.