Fix attachment of decoration to spec constants
If a phi instruction uses result of __spirv_SpecConst() call which has
not been translated yet, we created a OpForward instruction. Then we
translated the call to OpSpecConstant and attached the SpecId decoration
to it. After that, when we map LLVM value to SPIR-V values we detected
that there is a OpForward instruction and replaced Id of OpSpecConstant
with Id of the OpForward instruction, leaving the decoration targeted to
non-existing instruction.

To avoid this incorrect behaviour we should attach decorations after
replacing OpForward instruction, i.e. in transDecoration method.

Signed-off-by: Alexey Sotkin <alexey.sotkin@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/de22867