Handle OpSpecConstantOp with Select
`OpSpecConstantOp` may be used outside of a basic block (for example
in a variable initializer).  Modify the `SPIRVSelect` constructor to
accept a `nullptr` basic block.

Use an IRBuilder when translating `OpSelect`, so that the resulting
select expression will be created as a constant expression (and
possibly constant-folded) where possible.

Update some tests as the SPIR-V to LLVM translation now folds selects
more aggressively due to the use of IRBuilder.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/be24f4e