Add "-fsycl-device-only" to ESIMD diagnostics tests
The reason for the change is that "-verify" mode does not work correctly
with new integration footer approach, because it seems to ignore line
markers in the input file.

Since ESIMD tests for diagnostic messages only check device APIs and
device compilation is the first step in our compilation flow, added
"-fsycl-device-only" to avoid launching the host compilation.

The positive side of this change is that now those tests should be a bit
quicker as they do less work. The downside would be that we don't check
for diagnostics coming out of host compilation.