[SYCL] Lower queue::wait() to piQueueFinish when possible
This patch changes the logic of queue::wait() from waiting on each
individual event in order of submission of their tasks to checking if
each event's task has been enqueued, waiting for those that haven't been
and calling piQueueFinish to take care of the rest.

Notable exceptions to this new behaviour are host queues, queues that
emulate out-of-order execution by creating multiple queues underneath
and host task events, which are run on host regardless of the queue they
are bound to.