cl_khr_fp64 and cl_khr_fp16 extensions are now connected to the cuda plugin.
This change ensures that all Reduction tests run for the cuda backend.
The cl_khr_fp64 extension is used for all cuda devices by default, since sm_XX wherer XX < 13 has been unsupported by the cuda driver since cuda 8.0.
The test fp16-with-unnamed-lambda.cpp has been deleted because it has a duplicate in the test suite (in the dir SYCL/Regression).
In both cases the triple is missing on the first line which needs to be added to the llvm-test-suite copy to avoid a test failure now that the test is not skipped for the cuda backend.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>