[SYCL][NFC]Lowers overhead of making plugin calls
This change lowers the overhead of plugin calls by using const
char * strings, rather than std::string.  This completely eliminates
the calls to new and delete to allocate the strings, and also
gets rid of the std:string construction/destruction.  This also
reduces the code size of the library by about 0.5 percent.

This change should be NFC in terms of user visible changes to
SYCL behavior.