Renamed has_atomic_add_float to has_atomic_add_float64.
Renamed has_atomic_add_float to has_atomic_add_float64, since the general usage that includes float32 is only expected to be temporary.
has_atomic_add_float64 is a pseudonym of IsReduOptForAtomic64Add.
Updated documentation describing the current temporary usage of fp32 within IsReduOptForAtomic64Add.
IsReduOptForFastFloatAtomicAdd has been renamed IsReduOptForAtomic64Add to distinguish that this boolean should only be used in the case that the device has the sycl2020 atomic64 aspect, consistent with the naming convention used in other functions that are specializations for the atomic64 aspect.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>