[SYCL] Limit 'kernel name marking' to ONLY unnamed lambda cases.
The library is responsible to replace a non-named kernel's name with the
type of the lambda.  However, it still allows the user to specify the
name if they'd like, and uses that as the name instead.

The current implementation ALWAYS marked the kernel functor as the name
of the kernel, but this isn't correct, it isn't the name of the kernel
unless the library makes it so.  This ends up causing user-named kernels
that have their names changed by the functor they pass in.

We can't mark the first template parameter, since it might be a user
provided name, and we can't mark the functor since the user might have
provided a name.

However, if they are both the same (after unqualification and
canonicalization) that means that the library has done such a
replacement, and thus needs to be marked.