[Driver][NFC] Update usage of array container used for triple aliases
This class does not own the underlying data, it is
expected to be used in situations where the data resides in some
other buffer, whose lifetime extends past that of the ArrayRef. For
this reason, it is not in general safe to store an ArrayRef.
"Source: https://llvm.org/doxygen/classllvm_1_1ArrayRef.html