[SYCL][PI][L0] Changes std:string to const char * for ZeCallCount
This changes to use plain const char * for the ZeCallCount map
rather than std:string.  This saves std:string construction
and destruction overhead when ZE_DEBUG=4 is being used. It also
saves about 2% code size for the shared library due to this
overhead being removed.

This change should not result in any noticeable change in the
plug-ins behavior.