[SYCL][PI][L0] Force reset of memcpy command-list. (#4001)
This change set forces a host synchronization on a memcpy's
command list fence, and the reset of that commandlist. Resetting
the command list has the effect of causing Level Zero to
free any resources the command list may be using, and by doing
this as early as possible helps remove any bookkeeping that level
zero may be doing on memory regions.