[Driver][SYCL] Differentiate host dependency link from regular host link (#4002)
When performing DPC++ compilation with static archives, there is additional
work performed that gathers the host dependency information so we can
determine what device code should be pulled in.  Add an additional type
that will allow us to make heads or tails over which link step is being
performed.

This is needed to be able to override any user /link [arg] input that the
user may specify for MSVC compilations.  We do not want to honor the -out:*
information there during the host dependency link phase, as it will
override our expected output file designation.