[SYCL][CUDA] cl_khr_fp16 extension connected to cuda PI.
This change ensures that the fp16 Reduction test case runs for the cuda backend.
The test fp16-with-unnamed-lambda.cpp has been deleted because it has a duplicate in the test suite (in the dir SYCL/Regression).
In both cases the triple is missing on the first line which needs to be added to the llvm-test-suite copy to avoid a test failure now that the test is not skipped for the cuda backend.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>