Correct behavior of the __pf_kernel_wrapper to maintain namedness.
In order to keep the hack working, we put the __pf_kernel_wrapper as a
template parameter to the new lambda function (so its mangled name
contains the thing ot have attribute promotion).

The intent of this patch is to keep the kernel 'unnamed' (meaning both
the functor type and kernel name are the same) if the user specifies it,
but keep it NOT named otherwise.  This makes sure that our diagnostics
remain correct, as will our int-headers.