[SYCL] [Tests] Do not copy device binary image mocks (#4023)
Simplified, the motivation behind this change is the following.
Employing std::vector to store device binary images makes them
be allocated in heap which can't belong to neither any of loaded
shared objects nor to the executable itself. Moving allocation to
in-place makes the images be allocated in the executable.