[Driver][SYCL] Restrict user -include file in final integration footer step (#4036)
When compiling with the integration footer, the host side of the compile
is performed twice.  Once for the preprocessing and again with a full
file containing the integration footer.  For any user -include file.h
files, we only want to include that during the preprocessing step and not
with the subsequent compile.