[SYCL] Enable memcpy in libdevice (#3879)
Signed-off-by: gejin <ge.jin@intel.com>

Currently, memcpy is supported in SYCL device code if "-fno-builtin" is not added. Compiler will convert it to "llvm.memcpy..." intrinsic and llvm-spirv will convert it to OpCopyMemorySized for jit.

However, if "-fno-builtin" is added, compiler will call libc function "memcpy" directly and it requires libdevice support. This PR support memcpy in libdevice, so developers can use it in SYCL device code no matter whether "-fno-builtin" is added or not.