[SYCL] Skip PiCudaTests when running other plugins
This fixes an issue where if the cuda plugin is enabled and lit tests
for another plugin are run, it will try to run PiCudaTests, and all of
them will assert since the device filter will be set to a non-cuda
plugin and the tests request a cuda device.

So instead of asserting this patch makes the PiCudaTests get marked as
skipped if there is no cuda device found.