[SYCL] Ignore extra command line arguments for syntax only tests
It is possible to pass extra flags to clang for lit tests with
the CMake variable `SYCL_CLANG_EXTRA_FLAGS`.

So in some cases we may end up with command line arguments that are not
used in syntax only tests, such as `-mcpu` (which is required for ROCm
AMD), which clang will warn against, and since a lot of these tests are
also checking for any warnings, it will trip them up.

So this patch simply disables the warnings for unused command line
arguments in syntax only lit tests that check for warnings.