[SYCL][ROCm] Fix context destruction on AMD
What this piece of code is doing is making the context being destroyed
"current" by pushing it, so that it can be synchronized, and then
popping it.

However on AMD the synchronization is not supported, nor necessary, so
these steps can be skipped. In addition this was causing issues because
on AMD it uses a single context under the hood so `hipCtxt == current`
is always true, which means that this piece of code was popping the
context without pushing it first which would return an error code.

Simply skipping this step on AMD should be fine.