[Driver] Add an action to represent llvm-foreach and its handling during binding.
The patch adds ForEachWrappingClass which allows to encapsulate
a set of actions whose underlying commands will be wrapped under a llvm-foreach commands.

During the binding phases, the action will trigger the emission of jobs outside the set,
record the Jobs state then triggers the emission of the jobs inside the set.
The newly added commands are stolen from the Jobs list and
the commands are re-emitted with the llvm-foreach wrapping.

Signed-off-by: Victor Lomuller <victor@codeplay.com>