[SYCL] Add basic verification for accessor type template parameters
When processing accessors, we use the template parameters to make
various assumptions. It is important to ensure these paramaters are
defined as expected, to prevent the compiler from crashing due to a
rogue library implementation. This patch adds basic diagnostics for
this purpose.

As a drive-by, the patch also corrects a bug in the compiler to set
kernel function as invalid if an error has been thrown while
checking accessors.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>