[SYCL] Omit TemplatedDecl from int-footer for a VarTemplateDecl
The hook we had always added the VarDecl when it was visited, however
when evaluating a VarTemplateDecl, Clang creates a VarDecl to represent
the declaration as a child of the VarTemplateDecl. We don't need to
capture it in our collection since the VarTemplateSpecializationDecl
will provide the info we need.

If we DON'T do this, we end up having 2 specializations for the
specialization_id, one of which is invalid.