[SYCL][libclc] Remangler tool for consistent mangling
Libclc mangling is either done manually or done based on the function
signatures if requested. However, since libclc uses OpenCL C, the
primitive types are of set sizes and signedness, eg. long must be 64
bits and char must be signed.

When a target disagrees on primitive types it may use incompatible
libclc functions due to inconsistent mangling between libclc and the
target.

This commit adds a tool for remangling libclc bitcode files such that
the mangled names match the primitive types for the target.
Additionally it creates aliases where appropriate, eg. if long is 32
bits on the target a function using long will have its mangling changed
to appear as if it is using long long instead, and  the old mangling
will be made an alias to an existing 32-bit integer version of the
function if such a function exists.

Signed-off-by: Steffen Larsen <steffen.larsen@codeplay.com>