[SYCL] Allow overriding plugin libraries (#4067)
Motivation for this change is to provide ability to replay PI traces: a fake plugin opens a file with trace recording and replies to PI calls with info, acquired from that file. This can be used for debugging SYCL applications runtime when target hardware is not available. Such design allows the fake plugin rely on the fact that the library is only loaded once, which makes such a plugin stateless.