[Driver][SYCL] Enable way to emit int-footer source to a specific dir
During an -fsycl enabled build, we use an integration footer during the
host portion of the compilation.  This requires to create a temporary
file that is based on the original source file and contains the generated
footer.

Some build systems recognize this additional file that is being generated
and is not part of any expected generated dependency.  Introduce an option
that allows the user to specify the location in which this new source
file is created so the build system can properly track and recognize its
creation.  The new option is -fsycl-footer-path=<path>