Simplify "[[sycl::uses_aspects()]]"
We are planning to change `[[sycl::requires()]]` in the SYCL 2020 spec
to omit the `has()` clause, which means the parameter to that attribute
will just be a list of aspects.  Make the same change to
`[[sycl::uses_aspects()]]` in this design document.

Also add a note that we plan to change the name of
`[[sycl::requires()]]` to `[[sycl::device_has()]]`.  We can update this
design document again once that change is adopted into the SYCL 2020
spec.