[Driver][SYCL] Enable adding of default device triple
Add the default device triple (spir64) when we encounter any incoming objects
or libraries that have been previously built with the spir64 target.  The use
of -fno-sycl-link-spirv has been added to disable this behavior.

Also add a supplemental option of -fsycl-add-spirv which will add the default
device triple (spir64) if it is not already being used via -fsycl-targets.