[SYCL][ROCm] Fix building libclc for AMD (#4199)
This file uses `cl_mem_fence_flags`, `CLK_GLOBAL_MEM_FENCE`,
`CLK_LOCAL_MEM_FENCE`, which are defined by `clc.h`.

It wasn't included because when this file was written the OpenCL headers
were implicitely included, this has changed so now including the header
is necessary. See the discussion on #4133 for more details.