[SYCL] Construct context from first device in unit tests (#4142)
This is because multiple devices in a context is not supported with AOT
compiling, which is always used for the CUDA and ROCm pluginis, and
constructing the context from the platform may result in multiple
devices in the context if there are multiple devices available on the
system.

These tests are run with a device filter so this will only happen
when there is multiple devices of the same type. Which also means that
it should always be fine to take the first device from the platform as
it should still be of the expected device type.