[Driver][SYCL] Adjust header inclusion for source + generated footer compile (#4228)
When we are compiling the generated temporary source file that contains the
integration header, we cannot rely on the location of the source as one
of the default locations to search for headers.  We add the known location
of the original source to resolve those search directories.

When adding this location to the compilation, add via -internal-isystem as
opposed to -I so we can be sure that user -I values will win.