[SYCL] Omit extensions (.a .o .out .lib .obj .exe) in output project directory name
For "-o file.xxx" where the option value has an extension, if the
extension is one of .a .o .out .lib .obj .exe, the output project
directory name will be file.proj which omits the extension. Otherwise
the output project directory name will be file.xxx.prj which keeps
the original extension.

Signed-off-by: Qichao Gu <qichao.gu@intel.com>