[SYCL] Fix builtins address space type
This patch fixes the `core/GroupAsyncCopy.cl` and `ocl/prefetch.cl` lit
tests in `libclc`.

These started failing with the introduction of the SYCL address spaces
in #3634.

The issue is that the builtins are declared with the SYCL address spaces
in the tablegen so when used in an OpenCL setting it ends up with
address space mismatches between things like `opencl_global` and
`sycl_global`.

These two were the only tests affected because they're also the only
tested builtins that do not have generic address space variants in
tablegen, the other tests would just fall back on the generic address
space variants.

With this patch the builtins should automatically get the appropriate
version of the address space depending on the context.