Fix SPV-IR for OpBuildNDRange (#1145)
It is translated to a function with unmangled name
__spirv_BuildNDRange_{1|2|3}D with struct return parameter and array
arguments, since translator only translates it properly to SPIR-V with
this signature. _ND postfix is requred because array arguments are mangled
in the same way, so if there was no postfix, translator would produce
functions with same name for different dimensions.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/a6ca745