[SYCL] Correctly form the integration header for non-base-ascii char (#4287)
Unnamed kernels use the mangled name, which in some cases can contain an
arbitrary 'char' byte when unpacked using the integer_seq in the header.

Because identifiers can be a pretty massive set of unicode characters,
make sure that we can encode them properly.  We do this by being really
conservative and limiting ourselves to only printing as a
single-quote-delimited character literal when it is one of the
base-ascii printable characters, otherwise emit it as a signed integer
literal instead.