[SYCL][L0] Always set CommandList in _pi_context::getAvailableCommandList
When adding a command list from the cache to the map, we were not returning
the newly-inserted element via `CommandList` reference.

Thus, the _pi_context::getAvailableCommandList function was returning
`PI_SUCCESS`, but `CommandList` was left uninitialized, leading to a
segfault when the command list is used later (in my case, in `enqueueMemFillHelper`).

Introduced in PR #4248.