[SYCL] Create kernel bundle when non-native specialization constants are unset
The kernel bundle carrying the device memory used for specialization
constants on systems that do not support specialization constants
natively was only initialized if a specialization constant was either
written to or read from prior to kernel execution. These changes either
forces the creation of a kernel bundle if the kernel has a
specialization constant argument and a kernel bundle does not exist
already.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>