[SYCL] Cache kernel_id and implement sycl::get_kernel_ids
These changes implement the sycl::get_kernel_ids free function and
introduces caching of kernel identifiers in the program manager. The
caching of kernel identifiers is intended to ensure that each
identifier is only created once and copies of the identifier is passed
around, as to ensure equality in accordance with the SYCL 2020
specification.

NOTE: This is a non-breaking ABI change.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>