[ESIMD] Apply CRTP to simd_view_impl class (#4351)
CRTP - Curiously Recurring Template Pattern.

The main goal of the change is that it would not be possible
to generate objects of simd_view_impl class. To achieve
that we need to force simd_view_impl APIs to return objects
of derived class.

This change is needed for implementing writeable subscript
operator in simd_view.