seperate linking of kernel-bundles.
when fsycl-device-code-split=per_kernel is used in conjunction with multiple kernels, then explicit sycl::link(sycl::compile(my-kernel-bundle)) will fail, because we link all device images in a kernel together. But, the device-images are independent, there is presently no known situation where they should be linked together. So this PR works around this limitation by linking them seperately. This may have to be revisited once sharedlibrary style linking is supported, but that is likely true whether this bug is fixed or not.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>