[Driver][SYCL] Update dependency file generation with the integration footer
When generating dependency information and compiling the source + footer
file, the generated dependency information did not match the actual source
file that needs to be checked.

Update the compilation behavior when we are generating dependency information
with the source + footer so we do an additional specific dependency creation
step (preprocess only) against the original source file which is used instead
of the source + footer compilation.